x and y type corrected

+++ src/main/java/com/generation/localpro/dto/PortalUserResponseDTO.java
@@ -19,7 +19,7 @@ public class PortalUserResponseDTO {
     private String address;
     private List<String> roles;
     private String bio;
-    private int x;
-    private int y;
+    private Double x;
+    private Double y;
     private boolean isBanned;
 }
